Remove duplication of inserted fields
All the inserted fields are duplicated : they are converted to StringField, then added to the treeMap.
We can remove a lot of allocations by avoiding this behavior. Unless there is a specific reason for this purpose, it seems to works fine without this.
Furthermore, the setGroupCount() method allocated the string "1" when there is only one group, this is not efficient.