Comments on unguarded access to stopHandlingMessages in SocketInitiator
These comments are specifically related to the test

> SingleThreadedEventHandlingStrategyTest.shouldCleanUpInitiatorQFJMessageProcessorThreadAfterStop:248 : TestTimedOut

The call to stopHandlingMessages should happen after calling logout. Otherwise a logout message won't be sent.

```java
logoutAllSessions(forceDisconnect);
stopInitiators();
eventHandlingStrategy.stopHandlingMessages();
```

Assuming sessions have connected, this refactor is fine. When sessions do not connect as per this test, the `initialize` method (which ends up waiting for a connection) blocks the eventual call to `stop`, since both are mostly guarded by `this`.