Improve handling of Weekday schedules when EndTime < StartTime
In FX session typically starts on Sun evening and is reset around 5PM NewYork time. So daily schedule may look like this
 Sun 17:05 - Mon 16:55
 Mon 17:05 - Tue 16:55
 Tue 17:05 - Wed 16:55
 Wed 17:05 - Thu 16:55
 Thu 17:05 - Fri 16:55

I was expecting that below settings would achieve this:
```
  Weekdays = SUNDAY,MONDAY,TUESDAY,WEDNESDAY,THURSDAY,FRIDAY
  StartTime = 17:05:00
  EndTime = 16:55:00
```
But they don't, because DefaultSessionSchedule considers the interval [Fri 17:05 - Sat 16:55] to be in session - even if Sat is not a weekday.

This change improves weekday behaviour by going backwards until both start and end days are weekdays.
It does not have any impact in cases when EndTime > StartTime.

As a precaution, if suitable interval could not be found we revert to original behaviour. This should be configuration error, but even then returning something which is not quite right is better than getting stuck in infinite loop.