Proxy-Authorization header solves Proxy 407
HTTP Proxy authentication was failing with a 407 error code.
After an in-depth review, we discovered that the header sending was incorrect since the Proxy-Authorization key was missing.
This pull request aims to solve this issue by configuring the appropriate header for HttpProxyRequest whenever credentials are provided. Note that the code supports both NTLM and Basic authentication methods.
This code fixes our problem due to the correct authentication process from the headers set.