Adds exception DoNotFulfillResendRequest
Applications can throw this exception to abort the fulfillment of an
incoming ResendRequest.

Example use case: the counterparty sends an excessive number of
ResendRequests in error, and we want to prevent quickfix from fulfilling
them, spiking cpu and disk I/O, overwhelming the server.  Further action,
such as sending Rejects to alert the counterparty to the problem, are left
up to Applications.